Users can now cancel the new order form which takes them back to the last page
