+++ src/components/NewOrder.vue
@@ -4,8 +4,7 @@
       <router-link slot="menu" to="/">BUGSTERS</router-link>
     </MenuBar>
     <div class="content">
-      <a @click="goBack" class="btn">Go Back</a>
-      <OrderSection />
+      <OrderSection :showCancel="true" />
     </div>
 
     <FooterSection />
@@ -19,17 +18,10 @@ import FooterSection from './FooterSection'
 
 export default {
   name: 'NewOrder',
-
   components: {
     OrderSection,
     MenuBar,
     FooterSection
-  },
-
-  methods: {
-    goBack () {
-      this.$router.back()
-    }
   }
 }
 </script>
