+++ test/components/OrderForm.spec.js
@@ -1,3 +1,4 @@
+import VueRouter from 'vue-router'
 import VeeValidate, { Validator } from 'vee-validate'
 import { shallow, createLocalVue, mount } from 'vue-test-utils'
 
@@ -6,6 +7,9 @@ import OrderForm from '@/components/OrderForm'
 const localVue = createLocalVue()
 
 localVue.use(VeeValidate)
+localVue.use(VueRouter)
+
+const $router = new VueRouter()
 
 describe('OrderForm component', () => {
   it('should be a vue instance', () => {
@@ -27,7 +31,7 @@ describe('OrderForm component', () => {
       }
     })
 
-    let spy = jest.spyOn(wrapper.vm, 'handleSubmit')
+    const spy = jest.spyOn(wrapper.vm, 'handleSubmit')
 
     const submitButton = wrapper.find('button[type=submit]')
     submitButton.trigger('click')
@@ -42,7 +46,8 @@ describe('OrderForm component', () => {
         $validator: new Validator()
       }
     })
-    let spy = jest.spyOn(wrapper.vm, '$emit')
+
+    const spy = jest.spyOn(wrapper.vm, '$emit')
 
     const orderDetails = {
       budget: 500,
@@ -69,7 +74,7 @@ describe('OrderForm component', () => {
       }
     })
 
-    let spy = jest.spyOn(wrapper.vm, '$emit')
+    const spy = jest.spyOn(wrapper.vm, '$emit')
 
     const invalidOrderDetails = {
       budget: 1200,
@@ -85,6 +90,24 @@ describe('OrderForm component', () => {
     await wrapper.vm.handleSubmit('order-form')
 
     expect(spy).not.toHaveBeenCalled()
+  })
+
+  it('should navigate back when cancel button clicked', () => {
+    const wrapper = shallow(OrderForm, {
+      localVue,
+      router: $router,
+      provide: {
+        $validator: new Validator()
+      }
+    })
+
+    wrapper.setProps({ showCancel: true })
+
+    const spy = jest.spyOn(wrapper.vm.$router, 'back')
+
+    wrapper.vm.goBack()
+
+    expect(spy).toHaveBeenCalled()
   })
 
   it('snapshot', () => {
@@ -107,6 +130,21 @@ describe('OrderForm component', () => {
     })
 
     wrapper.setData({ hasError: true })
+
+    const $html = wrapper.vm.$el.outerHTML
+    expect($html).toMatchSnapshot()
+  })
+
+  it('snapshot with cancel button', () => {
+    const wrapper = mount(OrderForm, {
+      localVue,
+      provide: {
+        $validator: new Validator()
+      },
+      propsData: {
+        showCancel: true
+      }
+    })
 
     const $html = wrapper.vm.$el.outerHTML
     expect($html).toMatchSnapshot()
